likenft: Add ownerGrantRole/ownerRevokeRole for owner to modify the ACL
We opt to not using the adminRole feature to simplier ACL and avoid
transferring owner will not transfer the adminRole.

It will complicate the transfer of owner as it involve explicitly revoke
and renounce previous owner permission regarding update/mint. It will
preferable than old owner have access to owner permission.

refs likecoin/likecoin-op#121